search/requirements.txt
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.9.4
numpy==1.16.0
oauth2client==4.1.3
pandas==0.23.4
redis==3.0.1
rq==0.13.0
rq-dashboard==0.3.12
supervisor==4.0.3